move dev launcher auth to its own native module
