+++ packages/expo-dev-launcher/bundle/native-modules/DevLauncherInternal.ts
@@ -1,11 +1,4 @@
-import {
-  NativeModules,
-  NativeEventEmitter,
-  EventSubscription,
-  AppState,
-  Linking,
-  Platform,
-} from 'react-native';
+import { NativeModules, NativeEventEmitter, EventSubscription } from 'react-native';
 
 const DevLauncher = NativeModules.EXDevLauncherInternal;
 const EventEmitter = new NativeEventEmitter(DevLauncher);
@@ -51,117 +44,3 @@ export async function copyToClipboardAsync(content: string): Promise<null> {
 export const clientUrlScheme = DevLauncher.clientUrlScheme;
 export const installationID = DevLauncher.installationID;
 export const isDevice = !!DevLauncher.isDevice;
-
-let redirectHandler: ((event: any) => void) | null = null;
-let onWebBrowserCloseAndroid: null | (() => void) = null;
-let isAppStateAvailable: boolean = AppState.currentState !== null;
-
-function onAppStateChangeAndroid(state: any) {
-  if (!isAppStateAvailable) {
-    isAppStateAvailable = true;
-    return;
-  }
-
-  if (state === 'active' && onWebBrowserCloseAndroid) {
-    onWebBrowserCloseAndroid();
-  }
-}
-
-function stopWaitingForRedirect() {
-  if (!redirectHandler) {
-    throw new Error(
-      `The WebBrowser auth session is in an invalid state with no redirect handler when one should be set`
-    );
-  }
-
-  Linking.removeEventListener('url', redirectHandler);
-  redirectHandler = null;
-}
-
-async function openBrowserAndWaitAndroidAsync(startUrl: string): Promise<any> {
-  const appStateChangedToActive = new Promise<void>((resolve) => {
-    onWebBrowserCloseAndroid = resolve;
-    AppState.addEventListener('change', onAppStateChangeAndroid);
-  });
-
-  let result = { type: 'cancel' };
-  await DevLauncher.openWebBrowserAsync(startUrl);
-  const type = 'opened';
-
-  if (type === 'opened') {
-    await appStateChangedToActive;
-    result = { type: 'dismiss' };
-  }
-
-  AppState.removeEventListener('change', onAppStateChangeAndroid);
-  onWebBrowserCloseAndroid = null;
-  return result;
-}
-
-function waitForRedirectAsync(returnUrl: string): Promise<any> {
-  return new Promise((resolve) => {
-    redirectHandler = (event: any) => {
-      if (event.url.startsWith(returnUrl)) {
-        resolve({ url: event.url, type: 'success' });
-      }
-    };
-
-    Linking.addEventListener('url', redirectHandler);
-  });
-}
-
-async function openAuthSessionPolyfillAsync(startUrl: string, returnUrl: string): Promise<any> {
-  if (redirectHandler) {
-    throw new Error(
-      `The WebBrowser's auth session is in an invalid state with a redirect handler set when it should not be`
-    );
-  }
-
-  if (onWebBrowserCloseAndroid) {
-    throw new Error(`WebBrowser is already open, only one can be open at a time`);
-  }
-
-  try {
-    return await Promise.race([
-      openBrowserAndWaitAndroidAsync(startUrl),
-      waitForRedirectAsync(returnUrl),
-    ]);
-  } finally {
-    stopWaitingForRedirect();
-  }
-}
-
-export async function openAuthSessionAsync(url: string, returnUrl: string): Promise<any> {
-  if (DevLauncher.openAuthSessionAsync) {
-    // iOS
-    return await DevLauncher.openAuthSessionAsync(url, returnUrl);
-  }
-  // Android
-  return await openAuthSessionPolyfillAsync(url, returnUrl);
-}
-
-export async function getAuthSchemeAsync(): Promise<string> {
-  if (Platform.OS === 'android') {
-    return 'expo-dev-launcher';
-  }
-
-  return await DevLauncher.getAuthSchemeAsync();
-}
-
-export async function setSessionAsync(session: string): Promise<void> {
-  return await DevLauncher.setSessionAsync(session);
-}
-
-export async function restoreSessionAsync(): Promise<{
-  [key: string]: any;
-  sessionSecret: string;
-}> {
-  if (Platform.OS === 'android') {
-    try {
-      return JSON.parse(await DevLauncher.restoreSessionAsync());
-    } catch (exception) {
-      return null;
-    }
-  }
-  return await DevLauncher.restoreSessionAsync();
-}
