+++ packages/expo-dev-launcher/bundle/native-modules/__mocks__/DevLauncherAuth.ts
@@ -1,0 +1,4 @@
+export const openAuthSessionAsync = jest.fn().mockResolvedValue('123');
+export const setSessionAsync = jest.fn().mockResolvedValue(null);
+export const restoreSessionAsync = jest.fn().mockResolvedValue(null);
+export const getAuthSchemeAsync = jest.fn().mockResolvedValue('123');
