+++ packages/expo-dev-launcher/bundle/native-modules/__mocks__/DevLauncherInternal.ts
@@ -12,7 +12,3 @@ export const getPendingDeepLink = jest.fn().mockResolvedValue('');
 export const addDeepLinkListener = jest.fn().mockImplementation(() => jest.fn());
 export const copyToClipboardAsync = jest.fn().mockResolvedValue(null);
 export const getRecentlyOpenedApps = jest.fn().mockResolvedValue([]);
-export const openAuthSessionAsync = jest.fn().mockResolvedValue('123');
-export const setSessionAsync = jest.fn().mockResolvedValue(null);
-export const restoreSessionAsync = jest.fn().mockResolvedValue(null);
-export const getAuthSchemeAsync = jest.fn().mockResolvedValue('123');
