+++ packages/expo-dev-launcher/ios/EXDevLauncherAuth.h
@@ -1,0 +1,17 @@
+//
+//  EXDevLauncherAuth.h
+//  Pods
+//
+//  Created by andrew on 2022-02-02.
+//
+
+#import <Foundation/Foundation.h>
+#import <React/RCTBridgeModule.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface EXDevLauncherAuth : NSObject <RCTBridgeModule>
+
+@end
+
+NS_ASSUME_NONNULL_END
